+++ src/reporters/default.ts
@@ -1,6 +1,6 @@
 import { Reporter } from './Reporter'
 import { isLeft } from 'fp-ts/lib/Either'
-import { Context, getFunctionName } from '../index'
+import { Context, getFunctionName, ValidationError } from '../index'
 
 function stringify(value: any): string {
   return typeof value === 'function' ? getFunctionName(value) : JSON.stringify(value)
@@ -10,9 +10,17 @@ function getContextPath(context: Context): string {
   return context.map(({ key, type }) => `${key}: ${type.name}`).join('/')
 }
 
+function getMessage(value: any, context: Context): string {
+  return `Invalid value ${stringify(value)} supplied to ${getContextPath(context)}`
+}
+
+export function pathReporterFailure(es: Array<ValidationError>): Array<string> {
+  return es.map(e => getMessage(e.value, e.context))
+}
+
 export const PathReporter: Reporter<Array<string>> = {
   report: validation => validation.fold(
-    es => es.map(e => `Invalid value ${stringify(e.value)} supplied to ${getContextPath(e.context)}`),
+    pathReporterFailure,
     () => ['No errors!'],
   )
 }