+++ test/recursion.ts
@@ -10,25 +10,25 @@ describe('recursion', () => {
   it('should succeed validating a valid value', () => {
     const T = t.recursion('T', self => t.interface({
       a: t.number,
-      b: t.maybe(self)
+      b: t.union([self, t.undefined, t.null])
     }))
-    assertSuccess(t.validate({ a: 1 }, T))
-    assertSuccess(t.validate({ a: 1, b: { a: 2 } }, T))
+    assertSuccess(t.validate({ a: 1, b: null }, T))
+    assertSuccess(t.validate({ a: 1, b: { a: 2, b: null } }, T))
   })
 
   it('should return the same reference if validation succeeded', () => {
     const T = t.recursion('T', self => t.interface({
       a: t.number,
-      b: t.maybe(self)
+      b: t.union([self, t.undefined, t.null])
     }))
-    const value = { a: 1, b: { a: 2 } }
+    const value = { a: 1, b: { a: 2, b: null } }
     assertStrictEqual(t.validate(value, T), value)
   })
 
   it('should fail validating an invalid value', () => {
     const T = t.recursion('T', self => t.interface({
       a: t.number,
-      b: t.maybe(self)
+      b: t.union([self, t.undefined, t.null])
     }))
     assertFailure(t.validate(1, T), [
       'Invalid value 1 supplied to : T'
@@ -37,7 +37,7 @@ describe('recursion', () => {
       'Invalid value undefined supplied to : T/a: number'
     ])
     assertFailure(t.validate({ a: 1, b: {} }, T), [
-      'Invalid value undefined supplied to : T/b: ?T/a: number'
+      'Invalid value {} supplied to : T/b: (T | undefined | null)'
     ])
   })